fixes
Fixed problem with earlier tasks, optimazed code
