+++ src/styles/index.css
@@ -509,3 +509,11 @@ button {
 .google-signOut:hover {
   background-color: #4280fd;
 }
+
+.mb-4 {
+  margin-bottom: 4px;
+}
+
+.mb-6 {
+  margin-bottom: 6px;
+}
